+++ CHANGES.md
@@ -8,6 +8,7 @@
 - Shells no longer set their image on Mac, so that they don't hijack the dock icon.
 - Added `ControlWrapper.transparent` for efficiently passing `Control` to `ControlWrapper`.
 - Improved `StructuredDrop.handler`.
+- Added `SwtMisc.globalBounds` for `ControlWrapper`.
 
 ### Version 3.2.1 - August 7th 2019 ([javadoc](http://diffplug.github.io/durian-swt/javadoc/3.2.1/), [jcenter](https://bintray.com/diffplug/opensource/durian-swt/3.2.1/view))